pyproject.toml
[tool.poetry]
name = "hierarchical_data"
version = "0.1.0"
description = ""
authors = ["Mike Boring <shekinah999@msn.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
